fix: explicit uint32 conversion so it works in linux-386 platform build

+++ tcplisten/tcplisten.go
@@ -184,7 +184,7 @@ func safeIntToUint32(i int) (uint32, error) {
 	if i < 0 {
 		return 0, errors.New("value is negative, cannot convert to uint32")
 	}
-	if i > math.MaxUint32 {
+	if uint32(i) > uint32(math.MaxUint32) {
 		return 0, errors.New("value exceeds uint32 max value")
 	}
 	return uint32(i), nil
